refactor: extract command dispatch helper
